// src/components/pricing-tables/PricingTableSmall/styles.module.scss.d.ts
export type Styles = {
  Block: string;
  check: string;
  featureCell: string;
  featureCell__contents: string;
  featureCell__title: string;
  planCell: string;
  planCell__name: string;
  planCell__pricePeriod: string;
  planCell__priceValue: string;
  section: string;
  topHeader: string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;